Out of UTF-8 range now throws Brain-Flak not ruby Error
