+++ BrainFlakError.rb
@@ -1,0 +1,10 @@
+class BrainFlakError < StandardError
+
+  attr_reader :cause, :pos
+
+  def initialize(cause, pos)
+    @cause = cause
+    @pos = pos
+    super("Error at character %d: %s" % [pos, cause])
+  end
+end
